move api routes to top level
