+++ www/apps/api-reference/vercel.json
@@ -1,6 +1,6 @@
 {
   "crons": [{
-    "path": "/api/api/algolia",
+    "path": "/api/algolia",
     "schedule": "0 0 * * 4"
   }],
   "framework": "nextjs",
